+++ lib/internationalize/model.rb
@@ -348,6 +348,12 @@ module Internationalize
 
         Internationalize.locales.each do |locale|
           locale_str = locale.to_s
+
+          if locale_str.include?("-")
+            raise ArgumentError, "Locale '#{locale}' contains a hyphen which is invalid for Ruby method names. " \
+              "Use underscore format instead: :#{locale_str.tr('-', '_')}"
+          end
+
           getter_method = :"#{attr}_#{locale}"
 
           # Getter: article.title_en