+++ lib/internationalize/rich_text.rb
@@ -26,6 +26,15 @@ module Internationalize
       # @param name [Symbol] the attribute name
       #
       def international_rich_text(name)
+        # Validate locales don't contain hyphens (invalid for Ruby method names)
+        Internationalize.locales.each do |locale|
+          locale_str = locale.to_s
+          if locale_str.include?("-")
+            raise ArgumentError, "Locale '#{locale}' contains a hyphen which is invalid for Ruby method names. " \
+              "Use underscore format instead: :#{locale_str.tr('-', '_')}"
+          end
+        end
+
         # Generate has_rich_text for each locale
         Internationalize.locales.each do |locale|
           rich_text_name = :"#{name}_#{locale}"